requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.1
Flask-Compress==1.11
fonttools==4.31.2
future==0.18.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
plotly==5.7.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
python-twitter==3.5
pytz==2022.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
tenacity==8.0.1
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.9
vaderSentiment==3.3.2
Werkzeug==2.1.1
wordcloud==1.8.1
zipp==3.8.0
